+++ examples/resources/iosxe_interface_vrrp_v2/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxe_interface_vrrp_v2.example "<type>,<name>,<group_id>"